Factored Chunk back to input module...
